Do not add empty commit for portage repos (bug #201)
We currently setup /usr/portage and /../portage-overlay as empty repos
with a single commit which is required to branch. This leads to a merge
on the first pull which is unnecessary and requires manual intervention.

The solution is to not commit anything and use --orphan to create a
truly empty branch (when not using the master branch).